refactor(gym-settings): tabbed Details + Members; consolidate /members page
Surfaced from a UX consistency check on PR #135: the Members page lived at
its own /members route while invitations lived under /gym-settings. Two
related workflows in two places. This consolidates them.

UI
- /gym-settings is now a tabbed page: Details (default) and Members.
- The Details tab keeps the existing gym-name + timezone form.
- The new Members tab combines the old /members content (members table
  with role select, remove, program assignments) and the
  GymInvitationsPanel (invite-by-email form, pending list, resolved
  collapsible). Single place to manage gym membership lifecycle.
- Members table heading gains a count chip ('Members 12').
- Tab state mirrors window.location.hash (#members deep-links to the
  Members tab; back/forward and external links keep the right tab).
- Sidebar drops the standalone "Members" staff link — it's now a tab.
- /members now redirects to /gym-settings#members so old bookmarks land
  on the right tab.
- Members.tsx (the standalone page) is deleted; logic lives in the new
  components/MembersTab.tsx.

Programs and Pending Movements still hang out below the tabs; #139 tracks
moving them to dedicated places (ProgramsIndex / a reviewer-only route).

Tests
- tests/role-sidebar.spec.ts: STAFF_LINKS no longer expects 'Members' as a
  sidebar link (it's a tab inside Gym Settings).
- tests/gym-invitations.spec.ts: OWNER navigates to /gym-settings#members
  to reach the invite form.
- vitest 86/86, Playwright 29/29.
